+++ client/src/components/book/WarningBanner.tsx
